+++ rr/aaaa.js
@@ -3,7 +3,8 @@ const net = require('net')
 
 const sprintf = require('sprintf-js').sprintf
 
-const RR = require('./index').RR
+const RR      = require('./index').RR
+const TINYDNS = require('../lib/tinydns')
 
 class AAAA extends RR {
   constructor (opts) {
@@ -13,6 +14,7 @@ class AAAA extends RR {
     this.setAddress(opts?.address)
   }
 
+  /****** Resource record specific setters   *******/
   setAddress (val) {
     if (!val) throw new Error('AAAA: address is required')
     if (!net.isIPv6(val)) throw new Error('AAAA: address must be IPv6')
@@ -24,8 +26,21 @@ class AAAA extends RR {
     return [ 3596 ]
   }
 
+  /******  IMPORTERS   *******/
+  fromTinydns () {
+    // GENERIC =>  :fqdn:28:rdata:ttl:timestamp:lo
+    // AAAA       =>  3 fqdn : ip : x:ttl:timestamp:lo
+    // AAAA,PTR   =>  6 fqdn : ip : x:ttl:timestamp:lo
+  }
+
+  fromBind () {
+    //
+  }
+
+  /******  EXPORTERS   *******/
   toBind () {
-    return `${this.get('name')}\t${this.get('ttl')}\t${this.get('class')}\t${this.get('type')}\t${this.get('address')}\n`
+    const fields = [ 'name', 'ttl', 'class', 'type', 'address' ]
+    return `${fields.map(f => this.get(f)).join('\t')}\n`
   }
 
   toTinydns () {
@@ -42,11 +57,7 @@ class AAAA extends RR {
     val = val.split(':').map(s => sprintf('%04s', s)).join('')
 
     // from AAAA notation (8 groups of 4 hex digits) to 16 escaped octals
-    let rdata = ''
-    for (var i = 0; i < 32; i = i+2) {
-      // nibble off 2 hex bytes, encode to octal
-      rdata += sprintf('\\%03o', parseInt(val.slice(i, i+2), 16))
-    }
+    const rdata = TINYDNS.packHex(val)
 
     return `:${this.get('name')}:28:${rdata}:${this.getEmpty('ttl')}:${this.getEmpty('timestamp')}:${this.getEmpty('location')}\n`
   }