+++ rr/caa.js
@@ -1,5 +1,8 @@
 
+const sprintf = require('sprintf-js').sprintf
+
 const RR = require('./index').RR
+const TINYDNS = require('../lib/tinydns')
 
 class CAA extends RR {
   constructor (opts) {
@@ -7,10 +10,11 @@ class CAA extends RR {
     this.set('id', 257)
 
     this.setFlags(opts?.flags)
-    this.setTags(opts?.tags)
+    this.setTag(opts?.tag)
     this.setValue(opts?.value)
   }
 
+  /****** Resource record specific setters   *******/
   setFlags (val) {
     if (!this.is8bitInt('CAA', 'flags', val)) return
 
@@ -21,7 +25,7 @@ class CAA extends RR {
     this.set('flags', val)
   }
 
-  setTags (val) {
+  setTag (val) {
     if (typeof val !== 'string'
       || val.length < 1
       || /[A-Z]/.test(val)
@@ -29,9 +33,9 @@ class CAA extends RR {
       throw new Error('CAA flags must be a sequence of ASCII letters and numbers in lowercase: RFC 8659')
 
     if (![ 'issue', 'issuewild', 'iodef' ].includes(val)) {
-      console.warn(`CAA tags ${val} not recognized: RFC 6844`)
+      console.warn(`CAA tag ${val} not recognized: RFC 6844`)
     }
-    this.set('tags', val)
+    this.set('tag', val)
   }
 
   setValue (val) {
@@ -51,10 +55,34 @@ class CAA extends RR {
     return [ 6844 ]
   }
 
+  /******  IMPORTERS   *******/
+  fromTinydns () {
+    //
+  }
+
+  fromBind () {
+    //
+  }
+
+  /******  EXPORTERS   *******/
   toBind () {
-    let val = this.get('value')
-    if (val[0] !== '"') val = `"${val}"` // add enclosing quotes
-    return `${this.get('name')}\t${this.get('ttl')}\t${this.get('class')}\tCAA\t${this.get('flags')}\t${this.get('tags')}\t${val}\n`
+    let value = this.get('value')
+    if (value[0] !== '"') value = `"${value}"` // add enclosing quotes
+
+    const fields = [ 'name', 'ttl', 'class', 'type', 'flags', 'tag' ]
+    return `${fields.map(f => this.get(f)).join('\t')}\t${value}\n`
+  }
+
+  toTinydns () {
+    let rdata = ''
+    rdata += sprintf('\\%03o', this.get('flags'))
+
+    rdata += sprintf('\\%03o', this.get('tag').length)
+    rdata += TINYDNS.escapeOct(/[\r\n\t:\\/]/, this.get('tag'))
+
+    rdata += TINYDNS.escapeOct(/[\r\n\t:\\/]/, this.get('value'))
+
+    return `:${this.get('name')}:257:${rdata}:${this.getEmpty('ttl')}:${this.getEmpty('timestamp')}:${this.getEmpty('location')}\n`
   }
 }