+++ rr/dname.js
@@ -1,6 +1,7 @@
 const net = require('net')
 
 const RR = require('./index').RR
+const TINYDNS = require('../lib/tinydns')
 
 class DNAME extends RR {
   constructor (opts) {
@@ -15,6 +16,7 @@ class DNAME extends RR {
     }
   }
 
+  /****** Resource record specific setters   *******/
   setTarget (val) {
     if (!val) throw new Error('DNAME: target is required')
 
@@ -30,8 +32,24 @@ class DNAME extends RR {
     return [ 2672, 6672 ]
   }
 
+  /******  IMPORTERS   *******/
+  fromTinydns () {
+    //
+  }
+
+  fromBind () {
+    //
+  }
+
+  /******  EXPORTERS   *******/
   toBind () {
-    return `${this.get('name')}\t${this.get('ttl')}\t${this.get('class')}\tDNAME\t${this.get('target')}\n`
+    const fields = [ 'name', 'ttl', 'class', 'type', 'target' ]
+    return `${fields.map(f => this.get(f)).join('\t')}\n`
+  }
+
+  toTinydns () {
+    const rdata = TINYDNS.packDomainName(this.get('target'))
+    return `:${this.get('name')}:39:${rdata}:${this.getEmpty('ttl')}:${this.getEmpty('timestamp')}:${this.getEmpty('location')}\n`
   }
 }