+++ rr/srv.js
@@ -1,8 +1,6 @@
 
 const net = require('net')
 
-const sprintf = require('sprintf-js').sprintf
-
 const RR = require('./index').RR
 const TINYDNS = require('../lib/tinydns')
 
@@ -23,6 +21,7 @@ class SRV extends RR {
     }
   }
 
+  /****** Resource record specific setters   *******/
   setPriority (val) {
     if (!this.is16bitInt('SRV', 'priority', val)) return
 
@@ -56,8 +55,19 @@ class SRV extends RR {
     return [ 2782 ]
   }
 
+  /******  IMPORTERS   *******/
+  fromTinydns () {
+    //
+  }
+
+  fromBind () {
+    //
+  }
+
+  /******  EXPORTERS   *******/
   toBind () {
-    return `${this.get('name')}\t${this.get('ttl')}\t${this.get('class')}\tSRV\t${this.get('priority')}\t${this.get('weight')}\t${this.get('port')}\t${this.get('target')}\n`
+    const fields = [ 'name', 'ttl', 'class', 'type', 'priority', 'weight', 'port', 'target' ]
+    return `${fields.map(f => this.get(f)).join('\t')}\n`
   }
 
   toTinydns () {
@@ -65,10 +75,7 @@ class SRV extends RR {
     let rdata = ''
 
     for (const e of [ 'priority', 'weight', 'port' ]) {
-      const pri = Buffer.alloc(2)
-      pri.writeUInt16BE(this.get(e))
-      rdata += sprintf('\\%03o', pri.readUInt8(0))
-      rdata += sprintf('\\%03o', pri.readUInt8(1))
+      rdata += TINYDNS.UInt16AsOctal(this.get(e))
     }
 
     rdata += TINYDNS.packDomainName(this.get('target'))