+++ rr/sshfp.js
@@ -1,5 +1,6 @@
 
-const RR = require('./index').RR
+const RR      = require('./index').RR
+const TINYDNS = require('../lib/tinydns')
 
 class SSHFP extends RR {
   constructor (opts) {
@@ -11,6 +12,7 @@ class SSHFP extends RR {
     this.setFingerprint(opts?.fingerprint)
   }
 
+  /****** Resource record specific setters   *******/
   setAlgorithm (val) {
     // 0: reserved; 1: RSA 2: DSA 3: ECDSA 4: Ed25519 6:Ed448
     if (!this.is8bitInt('SSHFP', 'algorithm', val)) return
@@ -33,8 +35,30 @@ class SSHFP extends RR {
     return [ 4255 ]
   }
 
+  /******  IMPORTERS   *******/
+  fromTinydns () {
+    //
+  }
+
+  fromBind () {
+    //
+  }
+
+  /******  EXPORTERS   *******/
   toBind () {
-    return `${this.get('name')}\t${this.get('ttl')}\t${this.get('algorithm')}\t${this.get('fptype')}\t${this.get('fingerprint')}\n`
+    const fields = [ 'name', 'ttl', 'algorithm', 'fptype', 'fingerprint' ]
+    return `${fields.map(f => this.get(f)).join('\t')}\n`
+  }
+
+  toTinydns () {
+    let rdata = ''
+
+    for (const e of [ 'algo', 'type' ]) {
+      rdata += TINYDNS.UInt16AsOctal(this.get(e))
+    }
+
+    rdata += TINYDNS.packHex(this.get('fingerprint'))
+    return `:${this.get('name')}:44:${rdata}:${this.getEmpty('ttl')}:${this.getEmpty('timestamp')}:${this.getEmpty('location')}\n`
   }
 }