+++ rr/uri.js
@@ -1,5 +1,6 @@
 
-const RR = require('./index').RR
+const RR      = require('./index').RR
+const TINYDNS = require('../lib/tinydns')
 
 class URI extends RR {
   constructor (opts) {
@@ -11,6 +12,7 @@ class URI extends RR {
     this.setTarget(opts?.target)
   }
 
+  /****** Resource record specific setters   *******/
   setPriority (val) {
     if (!this.is16bitInt('URI', 'priority', val)) return
 
@@ -29,12 +31,36 @@ class URI extends RR {
     this.set('target', val)
   }
 
+  /******  IMPORTERS   *******/
+  fromTinydns () {
+    //
+  }
+
+  fromBind () {
+    //
+  }
+
+  /******  MISC   *******/
   getRFCs () {
     return [ 7553 ]
   }
 
+  /******  EXPORTERS   *******/
   toBind () {
-    return `${this.get('name')}\t${this.get('ttl')}\t${this.get('class')}\tURI\t${this.get('priority')}\t${this.get('weight')}\t${this.get('target')}\n`
+    const fields = [ 'name', 'ttl', 'class', 'type', 'priority', 'weight' ] // 'target'
+    return `${fields.map(f => this.get(f)).join('\t')}\t"${this.get('target')}"\n`
+  }
+
+  toTinydns () {
+    const dataRe = new RegExp(/[\r\n\t:\\/]/, 'g')
+    let rdata = ''
+
+    for (const e of [ 'priority', 'weight' ]) {
+      rdata += TINYDNS.UInt16AsOctal(this.get(e))
+    }
+
+    rdata += TINYDNS.escapeOct(dataRe, this.get('target'))
+    return `:${this.get('name')}:256:${rdata}:${this.getEmpty('ttl')}:${this.getEmpty('timestamp')}:${this.getEmpty('location')}\n`
   }
 }