+++ test/caa.js
@@ -11,20 +11,32 @@ const validRecords = [
     name : 'ns1.example.com',
     type : 'CAA',
     flags: 0,
-    tags : 'issue',
+    tag  : 'issue',
     value: 'letsencrypt.org',
     ttl  : 3600,
     testR: `ns1.example.com\t3600\tIN\tCAA\t0\tissue\t"letsencrypt.org"\n`,
+    testT: ':ns1.example.com:257:\\000\\005issueletsencrypt.org:3600::\n',
   },
   {
     class: 'IN',
-    name : 'ns1.example.com',
+    name : 'ns2.example.com',
     type : 'CAA',
     flags: 0,
-    tags : 'issue',
+    tag  : 'issue',
     value: '"lets crypt.org"',
     ttl  : 3600,
-    testR: `ns1.example.com\t3600\tIN\tCAA\t0\tissue\t"lets crypt.org"\n`,
+    testR: `ns2.example.com\t3600\tIN\tCAA\t0\tissue\t"lets crypt.org"\n`,
+    testT: ':ns2.example.com:257:\\000\\005issue"lets crypt.org":3600::\n',
+  },
+  {
+    name : 'example.net',
+    type : 'CAA',
+    ttl  : 86400,
+    flags: 0,
+    tag  : 'issuewild',
+    value: 'letsencrypt.org',
+    testR: 'example.net\t86400\tIN\tCAA\t0\tissuewild\t"letsencrypt.org"\n',
+    testT: ':example.net:257:\\000\\011issuewildletsencrypt.org:86400::\n',
   },
 ]
 
@@ -34,7 +46,7 @@ const invalidRecords = [
     name : 'example.com',
     type : 'CAA',
     flags: 1,
-    tags : 'issue',
+    tag  : 'issue',
     value: 'lets encrypt.org', // spaces aren't allowed unless quoted
     ttl  : 3600,
   },
@@ -45,16 +57,16 @@ describe('CAA record', function () {
   base.invalid(CAA, invalidRecords)
 
   for (const val of validRecords) {
-    it('converts to BIND format', async function () {
+    it(`converts to BIND format ${val.name}`, async function () {
       const r = new CAA(val).toBind()
       if (process.env.DEBUG) console.dir(r)
       assert.strictEqual(r, val.testR)
     })
 
-    it.skip('converts to tinydns format', async function () {
+    it(`converts to tinydns format ${val.name}`, async function () {
       const r = new CAA(val).toTinydns()
       if (process.env.DEBUG) console.dir(r)
-      assert.strictEqual(r, '\n')
+      assert.strictEqual(r, val.testT)
     })
   }
 })